requirements.txt
discord.py
pymongo
pyyaml
psutil
requests
gitpython